:rotating_light: Use TestUtil graph generator in test

+++ src/test/java/com/jacketing/io/output/format/DotFileOutputTest.java
@@ -1,28 +1,21 @@
 package com.jacketing.io.output.format;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.google.common.collect.HashBiMap;
+import com.jacketing.TestUtil;
 import com.jacketing.algorithm.impl.structures.Task;
 import com.jacketing.algorithm.interfaces.structures.Schedule;
 import com.jacketing.algorithm.interfaces.util.ScheduleFactory;
 import com.jacketing.io.cli.ProgramContext;
 import com.jacketing.io.output.saver.StandardFileSaver;
-import com.jacketing.parsing.impl.Parser;
-import com.jacketing.parsing.impl.services.EnumerationService;
-import com.jacketing.parsing.impl.services.WeightService;
-import com.jacketing.parsing.impl.structures.EnumeratedAdjacencyList;
-import com.jacketing.parsing.impl.structures.Graph;
-import com.jacketing.parsing.interfaces.structures.services.EnumeratedNodeMap;
-import com.paypal.digraph.parser.GraphParser;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.HashMap;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -32,70 +25,20 @@ public class DotFileOutputTest {
   DotFileFormatter fileFormatter;
   StandardFileSaver fileSaver;
 
-  private static Graph createGraph(GraphParser graph) {
-    EnumerationService enumerationService = new EnumerationService(
-      HashBiMap.create()
-    );
-    EnumeratedNodeMap enumeratedNodeMap = enumerationService.enumerateFromGraph(
-      graph
-    );
-    return new Graph(
-      new EnumeratedAdjacencyList(
-        graph,
-        enumeratedNodeMap,
-        new HashMap<>(),
-        new HashMap<>()
-      ),
-      new WeightService(
-        graph,
-        enumeratedNodeMap,
-        new HashMap<>(),
-        new HashMap<>()
-      )
-    );
-  }
-
-  private static Graph createGraphFromBuffer(String graphString) {
-    try {
-      com.jacketing.parsing.impl.Parser parser = Parser.fromStringBuffer(
-        new StringBuffer(graphString)
-      );
-      GraphParser graph = parser.parse();
-      return createGraph(graph);
-    } catch (Exception e) {
-      return null;
-    }
-  }
-
   @Before
-  public void setUp() throws Exception {
+  public void setUp() {
     fileFormatter = new DotFileFormatter();
     fileSaver = new StandardFileSaver();
   }
 
   @After
-  public void tearDown() throws Exception {
+  public void tearDown() {
     File fileToDelete = new File("output");
-    if (!fileToDelete.delete()) {
-      System.out.println("File output was not deleted.");
-    } else {
-      System.out.println("File output deleted.");
+    if (fileToDelete.exists()) {
+      if (!fileToDelete.delete()) {
+        fail("Warning: Failed to delete the test-generated file 'output'");
+      }
     }
-  }
-
-  public Graph getGraphG1() {
-    return createGraphFromBuffer(
-      "digraph \"g1\" {\n" +
-      "  a [Weight = 2];\n" +
-      "  b [Weight = 3];\n" +
-      "  a -> b [Weight = 1];\n" +
-      "  c [Weight = 3];\n" +
-      "  a -> c [Weight = 2];\n" +
-      "  d [Weight = 2];\n" +
-      "  b -> d [Weight = 2];\n" +
-      "  c -> d [Weight = 1];\n" +
-      "}\n"
-    );
   }
 
   public Schedule getScheduleG1() {
@@ -127,7 +70,10 @@ public class DotFileOutputTest {
 
   @Test
   public void checkOutputCorrect() {
-    String output = fileFormatter.formatSchedule(getScheduleG1(), getGraphG1());
+    String output = fileFormatter.formatSchedule(
+      getScheduleG1(),
+      TestUtil.graphVariantOne()
+    );
     assertEquals(getValidOutputString(), output);
   }
 
@@ -139,6 +85,5 @@ public class DotFileOutputTest {
       StandardCharsets.UTF_8
     );
     assertEquals(getValidOutputString(), content);
-    System.out.println("File output created");
   }
 }
